[D83T-42] 위도, 경도 자료형 수정

+++ src/main/java/d83t/bpmbackend/domain/aggregate/location/dto/LocationRequestDto.java
@@ -14,12 +14,12 @@ public class LocationRequestDto {
     @NotBlank
     private String address;
     @NotNull
-    private int latitude;
+    private double latitude;
     @NotNull
-    private int longitude;
+    private double longitude;
 
     @Builder
-    public LocationRequestDto(String address, int latitude, int longitude) {
+    public LocationRequestDto(String address, double latitude, double longitude) {
         this.address = address;
         this.latitude = latitude;
         this.longitude = longitude;
